@@ -1,7 +1,7 @@
 [project]
 name = "AYON"
-version = "1.1.2-dev.1"
-requires-python = "<3.10"
+version = "1.3.4-dev"
+requires-python = "3.11"
 package-mode = false
 description = "Open VFX and Animation pipeline with support."
 authors = [
@@ -25,7 +25,7 @@ packages = [
 ]
 
 dependencies = [
-    "ayon-python-api",
+    "ayon-python-api ~= 1.1.3",
     "arrow ~= 0.17",
     "Unidecode ~= 1.2.0",
     "aiohttp ~= 3.7",
@@ -45,8 +45,11 @@ dependencies = [
     "urllib3 ~= 1.26.16",
     "semver ~= 2.13.0",
     "QtPy ~= 2.3.0",
+    "pywin32 == 301; sys_platform == 'win32'",
+    "python3-xlib; sys_platform == 'linux'",
+    "distro == 1.9.0; sys_platform == 'linux'",
     "pip ~= 24.0",
-    "pre-commit"
+
 ]
 
 [project.optional-dependencies]
@@ -56,13 +59,15 @@ dev = [
     "coverage",
     "cx_Freeze ~= 6.12.0",
     "GitPython ~= 3.1.17",
+    "hatchling",
     "jedi ~= 0.13",
-    "Jinja2",
+    "Jinja2 ~= 3",
     "markupsafe ~= 2.0.1",
     "pycodestyle",
     "pydocstyle",
     "linkify-it-py ~= 2.0.0",
     "myst-parser ~= 0.18.1",
+    "precommit",
     "pylint ~= 2.5",
     "pytest ~= 6.1",
     "pytest-cov",
