@@ -1,6 +1,7 @@
 import os
 import sys
 import json
+import time
 import uuid
 import ctypes
 import tempfile
@@ -9,12 +10,13 @@ import collections
 import datetime
 import logging
 import shutil
-import threading
 import platform
 import subprocess
+import dataclasses
 from enum import Enum
+from abc import ABC, abstractmethod
+from typing import Optional, Any, Generator
 
-import attr
 import ayon_api
 
 from ayon_common.utils import (
@@ -24,22 +26,75 @@ from ayon_common.utils import (
     is_dev_mode_enabled,
     get_executables_info_by_version,
     get_downloads_dir,
+    calculate_file_checksum,
 )
 
 from .exceptions import BundleNotFoundError, InstallerDistributionError
-from .utils import (
-    get_addons_dir,
-    get_dependencies_dir,
-)
-from .downloaders import get_default_download_factory
 from .data_structures import (
+    SourceInfo,
     Installer,
     AddonInfo,
     DependencyItem,
     Bundle,
 )
+from .utils import (
+    get_addons_dir,
+    get_dependencies_dir,
+)
+from .downloaders import (
+    SourceDownloader,
+    get_default_download_factory,
+    DownloadFactory,
+)
 
 NOT_SET = type("UNKNOWN", (), {"__bool__": lambda: False})()
+DIST_PROGRESS_FILENAME = "dist_progress.json"
+MOVE_WAIT_TRESHOLD_TIME = 20
+IS_WINDOWS = platform.system().lower() == "windows"
+
+
+class DistributionProgressInterupted(Exception):
+    pass
+
+
+def _windows_dir_requires_permissions(dirpath: str) -> bool:
+    try:
+        # Attempt to create a temporary file in the folder
+        temp_file_path = os.path.join(dirpath, uuid.uuid4().hex)
+        with open(temp_file_path, "w"):
+            pass
+        os.remove(temp_file_path)  # Clean up temporary file
+        return False
+
+    except PermissionError:
+        return True
+
+    except BaseException as exc:
+        print(
+            "Failed to determine if root requires permissions."
+            f"Unexpected error: {exc}"
+        )
+        return False
+
+
+def _has_write_permissions(dirpath: str) -> bool:
+    while not os.path.exists(dirpath):
+        _dirpath = os.path.dirname(dirpath)
+        if _dirpath == dirpath:
+            print((
+                "Failed to determine if root requires permissions."
+                " The disk is probably not mounted."
+            ))
+            return False
+        dirpath = _dirpath
+
+    if IS_WINDOWS:
+        if ctypes.windll.shell32.IsUserAnAdmin():
+            return True
+
+        return not _windows_dir_requires_permissions(dirpath)
+
+    return os.access(dirpath, os.W_OK)
 
 
 class UpdateState(Enum):
@@ -48,6 +103,241 @@ class UpdateState(Enum):
     OUTDATED = "outdated"
     UPDATE_FAILED = "failed"
     MISS_SOURCE_FILES = "miss_source_files"
+
+
+# --- Distribution metadata file ---
+# The file is stored to destination directory to keep track of distribution
+#   of the item.
+class DistFileStates(Enum):
+    moving = "moving"
+    done = "done"
+    failed = "failed"
+
+
+def _read_progress_file(progress_dir: str):
+    progress_path = os.path.join(progress_dir, DIST_PROGRESS_FILENAME)
+    try:
+        with open(progress_path, "r") as stream:
+            return json.loads(stream.read())
+    except Exception:
+        return {}
+
+
+def _store_progress_file(progress_dir: str, progress_info: dict[str, Any]):
+    progress_path = os.path.join(progress_dir, DIST_PROGRESS_FILENAME)
+    if not os.path.exists(progress_dir):
+        os.makedirs(progress_dir, exist_ok=True)
+    with open(progress_path, "w") as stream:
+        json.dump(progress_info, stream)
+
+
+def _create_progress_file(
+    progress_dir: str,
+    progress_id: str,
+    checksum: Optional[str],
+    checksum_algorithm: Optional[str],
+    state: Optional[DistFileStates] = None
+):
+    if state is None:
+        state = DistFileStates.moving
+
+    progress_info = {
+        # State of the progress
+        "state": state.value,
+        # Unique identifier of the progress
+        "progress_id": progress_id,
+        # Just a metadata about the file
+        "checksum": checksum,
+        "checksum_algorithm": checksum_algorithm,
+        "dist_started": (
+            datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        ),
+    }
+    _store_progress_file(progress_dir, progress_info)
+
+
+def _clean_dist_dir(dist_dirpath: str):
+    if not os.path.exists(dist_dirpath):
+        return
+    for subname in os.listdir(dist_dirpath):
+        if subname == DIST_PROGRESS_FILENAME:
+            continue
+        path = os.path.join(dist_dirpath, subname)
+        if os.path.isdir(path):
+            shutil.rmtree(path)
+        else:
+            os.remove(path)
+
+
+def _wait_for_other_process(
+    progress_dir: str,
+    progress_id: str,
+    log: logging.Logger,
+) -> Generator[Optional[bool], None, None]:
+    progress_path = os.path.join(progress_dir, DIST_PROGRESS_FILENAME)
+    started = time.time()
+    progress_existed = False
+    threshold_time = None
+    state = None
+    output = False
+    while True:
+        if not os.path.exists(progress_path):
+            if progress_existed:
+                log.debug(
+                    "Other processed didn't finish download or extraction,"
+                    " trying to do so."
+                )
+            break
+
+        progress_info = _read_progress_file(progress_dir)
+        if progress_info.get("progress_id") == progress_id:
+            break
+
+        current_state = progress_info.get("state")
+
+        if not progress_existed:
+            log.debug(
+                "Other process already created progress file"
+                " in target directory. Waiting for finishing it."
+            )
+
+        progress_existed = True
+        if current_state is None:
+            log.warning(
+                "Other process did not store 'state' to progress file."
+            )
+            break
+
+        if current_state == DistFileStates.done.value:
+            log.debug("Other process finished distribution.")
+            output = True
+            break
+
+        if current_state == DistFileStates.failed.value:
+            log.debug("Other process failed distribution.")
+            state = current_state
+            threshold_time = 0
+
+        if current_state != state:
+            started = time.time()
+            threshold_time = None
+            state = current_state
+
+        if threshold_time is None:
+            threshold_time = MOVE_WAIT_TRESHOLD_TIME
+
+        if (time.time() - started) > threshold_time:
+            log.debug(
+                f"Waited for treshold time ({threshold_time}s)."
+                f" Extracting downloaded content."
+            )
+            _clean_dist_dir(progress_dir)
+            break
+        yield None
+    yield output
+
+
+def change_permissions_recursive(path: str, mode: int = 0o777):
+    if IS_WINDOWS or not os.path.exists(path):
+        return
+
+    if os.path.isfile(path):
+        # If the path is a file, change its permissions directly
+        os.chmod(path, mode)
+        return
+
+    # Traverse the directory tree starting from 'path' to top
+    for root, dirnames, filenames in os.walk(path, topdown=False):
+        # Change the permissions of each directory
+        for dirname in dirnames:
+            dirpath = os.path.join(root, dirname)
+            os.chmod(dirpath, mode)
+
+        for filename in filenames:
+            filepath = os.path.join(root, filename)
+            os.chmod(filepath, mode)
+
+
+# --- Distribution download directories ---
+# Where to store downloaded files before extration.
+def _get_dist_download_dir(*args):
+    return os.path.join(
+        tempfile.gettempdir(), "ayon_dist_downloads", *args
+    )
+
+
+def _create_dist_expire_file(process_dir: str):
+    """Create distribution temp directory with metadata file.
+
+    The metadata file contains information about expiration time of the
+        temp folder. Lifetime is 1 hour (more than should be needed).
+
+    Args:
+        process_dir (str): Path to distribution temp directory.
+
+    """
+    info_path = os.path.join(process_dir, "download_info.json")
+    if os.path.exists(info_path):
+        return
+
+    if not os.path.exists(process_dir):
+        os.makedirs(process_dir, exist_ok=True)
+
+    with open(info_path, "w") as stream:
+        json.dump(
+            {"expiration_time": time.time() + (60 * 60)},
+            stream,
+        )
+
+    change_permissions_recursive(process_dir)
+
+
+def _dist_expire_file_expired(process_dir: str) -> bool:
+    """Check if distribution temp directory is expired.
+
+    Args:
+        process_dir (str): Path to distribution temp directory.
+
+    Returns:
+        bool: Directory is expired and can be removed.
+
+    """
+    info_path = os.path.join(process_dir, "download_info.json")
+    if not os.path.exists(info_path):
+        return True
+
+    try:
+        with open(info_path, "r") as stream:
+            data = json.load(stream)
+    except Exception:
+        data = {}
+    expiration_time = data.get("expiration_time")
+    if not isinstance(expiration_time, int):
+        return False
+    return expiration_time < time.time()
+
+
+def _cleanup_dist_expire_dirs(process_dir: str):
+    if not os.path.exists(process_dir):
+        return
+    for subname in os.listdir(process_dir):
+        path = os.path.join(process_dir, subname)
+        if os.path.isdir(path) and _dist_expire_file_expired(path):
+            try:
+                shutil.rmtree(path)
+            except Exception:
+                print(
+                    f"Failed to remove expired distribution directory: {path}"
+                )
+
+
+def _cleanup_dist_download_dirs():
+    """Clean up old distribution download directories.
+
+    If distribution crashed in past this function makes sure they are removed.
+
+    """
+    _cleanup_dist_expire_dirs(_get_dist_download_dir())
 
 
 class DistributeTransferProgress:
@@ -71,13 +361,13 @@ class DistributeTransferProgress:
 
         self._started = True
 
-    def set_failed(self, reason):
+    def set_failed(self, reason: str):
         """Set source distribution as failed.
 
         Args:
             reason (str): Error message why the transfer failed.
-        """
 
+        """
         self._failed = True
         self._fail_reason = reason
 
@@ -102,7 +392,7 @@ class DistributeTransferProgress:
         self._unzip_finished = True
 
     @property
-    def is_running(self):
+    def is_running(self) -> bool:
         """Source distribution is in progress.
 
         Returns:
@@ -116,45 +406,45 @@ class DistributeTransferProgress:
         )
 
     @property
-    def transfer_progress(self):
+    def transfer_progress(self) -> ayon_api.TransferProgress:
         """Source file 'download' progress tracker.
 
         Returns:
-            ayon_api.TransferProgress.: Content download progress.
-        """
+            ayon_api.TransferProgress: Content download progress.
 
+        """
         return self._transfer_progress
 
     @property
-    def started(self):
+    def started(self) -> bool:
         return self._started
 
     @property
-    def hash_check_started(self):
+    def hash_check_started(self) -> bool:
         return self._hash_check_started
 
     @property
-    def hash_check_finished(self):
+    def hash_check_finished(self) -> bool:
         return self._hash_check_finished
 
     @property
-    def unzip_started(self):
+    def unzip_started(self) -> bool:
         return self._unzip_started
 
     @property
-    def unzip_finished(self):
+    def unzip_finished(self) -> bool:
         return self._unzip_finished
 
     @property
-    def failed(self):
+    def failed(self) -> bool:
         return self._failed or self._transfer_progress.failed
 
     @property
-    def fail_reason(self):
+    def fail_reason(self) -> Optional[str]:
         return self._fail_reason or self._transfer_progress.fail_reason
 
 
-class BaseDistributionItem:
+class BaseDistributionItem(ABC):
     """Distribution item with sources and target directories.
 
     Distribution item can be an installer, addon or dependency package.
@@ -174,120 +464,139 @@ class BaseDistributionItem:
         download_dirpath (str): Path to directory where file is unzipped.
         state (UpdateState): Initial state (UpdateState.UPDATED or
             UpdateState.OUTDATED).
-        checksum (str): Hash of file for validation.
-        checksum_algorithm (str): Algorithm used to generate the hash.
+        checksum (Optional[str]): Hash of file for validation.
+        checksum_algorithm (Optional[str]): Algorithm used to generate the hash.
         factory (DownloadFactory): Downloaders factory object.
         sources (List[SourceInfo]): Possible sources to receive the
             distribution item.
         downloader_data (Dict[str, Any]): More information for downloaders.
         item_label (str): Label used in log outputs (and in UI).
-        logger (logging.Logger): Logger object.
-    """
+        progress_dir (Optional[str]): Directory where progress is stored for
+            other processes.
+        logger (Optional[logging.Logger]): Logger object.
 
+    """
     def __init__(
         self,
-        download_dirpath,
-        state,
-        checksum,
-        checksum_algorithm,
-        factory,
-        sources,
-        downloader_data,
-        item_label,
-        logger=None,
+        download_dirpath: str,
+        state: UpdateState,
+        checksum: Optional[str],
+        checksum_algorithm: Optional[str],
+        factory: DownloadFactory,
+        sources: list[SourceInfo],
+        downloader_data: dict[str, Any],
+        item_label: str,
+        progress_dir: Optional[str] = None,
+        logger: Optional[logging.Logger] = None,
     ):
         if logger is None:
             logger = logging.getLogger(self.__class__.__name__)
-        self.log = logger
-        self.state = state
-        self.download_dirpath = download_dirpath
-        self.checksum = checksum
-        self.checksum_algorithm = checksum_algorithm
-        self.factory = factory
+        self.log: logging.Logger = logger
+        self.state: UpdateState = state
+        self.download_dirpath: str = download_dirpath
+        self.checksum: Optional[str] = checksum
+        self.checksum_algorithm: Optional[str] = checksum_algorithm
+        self.factory: DownloadFactory = factory
         self.sources = self._prepare_sources(sources)
-        self.downloader_data = downloader_data
-        self.item_label = item_label
+        self.downloader_data: dict[str, Any] = downloader_data
+        self.item_label: str = item_label
 
         self._need_distribution = state != UpdateState.UPDATED
         self._current_source_progress = None
         self._used_source_progress = None
         self._used_source = None
         self._dist_started = False
-        self._dist_finished = False
+
+        self._progress_id = uuid.uuid4().hex
+        self._progress_dir = progress_dir
 
         self._error_msg = None
         self._error_detail = None
 
-    def _prepare_sources(self, sources):
+    @staticmethod
+    def _prepare_sources(sources: list[SourceInfo]):
         return [
             (source, DistributeTransferProgress())
             for source in sources
         ]
 
     @property
-    def need_distribution(self):
+    def need_distribution(self) -> bool:
         """Need distribution based on initial state.
 
         Returns:
             bool: Need distribution.
-        """
 
+        """
         return self._need_distribution
 
     @property
-    def current_source_progress(self):
+    def current_source_progress(self) -> Optional[DistributeTransferProgress]:
         """Currently processed source progress object.
 
         Returns:
-            Union[DistributeTransferProgress, None]: Transfer progress or None.
-        """
+            Optional[DistributeTransferProgress]: Transfer progress or None.
 
+        """
         return self._current_source_progress
 
     @property
-    def used_source_progress(self):
+    def used_source_progress(self) -> Optional[DistributeTransferProgress]:
         """Transfer progress that successfully distributed the item.
 
         Returns:
-            Union[DistributeTransferProgress, None]: Transfer progress or None.
-        """
+            Optional[DistributeTransferProgress]: Transfer progress or None.
 
+        """
         return self._used_source_progress
 
     @property
-    def used_source(self):
+    def used_source(self) -> Optional[dict[str, Any]]:
         """Data of source item.
 
         Returns:
-            Union[Dict[str, Any], None]: SourceInfo data or None.
-        """
+            Optional[dict[str, Any]]: SourceInfo data or None.
 
+        """
         return self._used_source
 
     @property
-    def error_message(self):
+    def error_message(self) -> Optional[str]:
         """Reason why distribution item failed.
 
         Returns:
-            Union[str, None]: Error message.
+            Optional[str]: Error message.
         """
 
         return self._error_msg
 
     @property
-    def error_detail(self):
+    def error_detail(self) -> Optional[str]:
         """Detailed reason why distribution item failed.
 
         Returns:
-            Union[str, None]: Detailed information (maybe traceback).
+            Optional[str]: Detailed information (maybe traceback).
         """
 
         return self._error_detail
 
-    def _pre_source_process(self):
-        os.makedirs(self.download_dirpath, exist_ok=True)
+    @property
+    @abstractmethod
+    def is_missing_permissions(self) -> bool:
+        pass
 
-    def _receive_file(self, source_data, source_progress, downloader):
+    def _pre_source_process(self):
+        if not os.path.exists(self.download_dirpath):
+            os.makedirs(self.download_dirpath, exist_ok=True)
+
+        change_permissions_recursive(self.download_dirpath)
+
+    def _receive_file(
+        self,
+        source_data: dict[str, Any],
+        source_progress: DistributeTransferProgress,
+        downloader: SourceDownloader,
+    ) -> Optional[str]:
         """Receive source filepath using source data and downloader.
 
         Args:
@@ -298,9 +607,9 @@ class BaseDistributionItem:
                 about receiving file from source.
 
         Returns:
-            Union[str, None]: Filepath to received file from source.
-        """
+            Optional[str]: Filepath to received file from source.
 
+        """
         download_dirpath = self.download_dirpath
 
         try:
@@ -328,6 +637,15 @@ class BaseDistributionItem:
                 downloader.check_hash(
                     filepath, self.checksum, self.checksum_algorithm
                 )
+            else:
+                # Fill checksum automatically based on downloaded file
+                # - still better than nothing
+                if not self.checksum_algorithm:
+                    self.checksum_algorithm = "sha256"
+                self.checksum = calculate_file_checksum(
+                    filepath, self.checksum_algorithm
+                )
+
         except Exception:
             message = "File hash does not match"
             source_progress.set_failed(message)
@@ -340,8 +658,12 @@ class BaseDistributionItem:
         return filepath
 
     def _post_source_process(
-        self, filepath, source_data, source_progress, downloader
-    ):
+        self,
+        filepath: str,
+        source_data: dict[str, Any],
+        source_progress: DistributeTransferProgress,
+        downloader: SourceDownloader,
+    ) -> Generator[Optional[bool], None, None]:
         """Process source after it is downloaded and validated.
 
         Override this method if downloaded file needs more logic to do, like
@@ -353,15 +675,17 @@ class BaseDistributionItem:
         Args:
             filepath (str): Path to a downloaded source.
             source_data (dict[str, Any]): Source information data.
+            source_progress (DistributeTransferProgress): Source progress.
             downloader (SourceDownloader): Object which cared about download
                 of file.
 
         Returns:
-            bool: Post processing finished in a way that it is not needed to
-                process other possible sources. Does not mean that it was
-                successful.
-        """
+            Generator[Optional[bool], None, None]: Post processing finished
+                in a way that it is not needed to process other possible
+                sources. Does not mean that it was successful. If yields None
+                then is still working.
 
+        """
         if filepath:
             self.state = UpdateState.UPDATED
             self._used_source = source_data
@@ -371,9 +695,13 @@ class BaseDistributionItem:
             self.download_dirpath,
             self.downloader_data
         )
-        return bool(filepath)
+        yield bool(filepath)
 
-    def _process_source(self, source, source_progress):
+    def _process_source(
+        self,
+        source: SourceInfo,
+        source_progress: DistributeTransferProgress,
+    ) -> Generator[Optional[bool], None, None]:
         """Process single source item.
 
         Cares about download, validate and process source.
@@ -381,13 +709,14 @@ class BaseDistributionItem:
         Args:
             source (SourceInfo): Source information.
             source_progress (DistributeTransferProgress): Object to keep track
-                about process of an source.
+                about process of a source.
 
         Returns:
-            bool: Source was processed so any other sources can be skipped.
-                Does not have to be successfull.
-        """
+            Generator[Optional[bool], None, None]: Source was processed so
+                any other sources can be skipped. Does not have to be
+                successfull. Is still working if yields None.
 
+        """
         self._current_source_progress = source_progress
         source_progress.set_started()
 
@@ -398,16 +727,18 @@ class BaseDistributionItem:
             message = f"Unknown downloader {source.type}"
             source_progress.set_failed(message)
             self.log.warning(message, exc_info=True)
-            return False
+            yield False
+            return
 
         try:
-            source_data = attr.asdict(source)
+            source_data = dataclasses.asdict(source)
             filepath = self._receive_file(
                 source_data,
                 source_progress,
                 downloader
             )
-            return self._post_source_process(
+
+            yield from self._post_source_process(
                 filepath, source_data, source_progress, downloader
             )
 
@@ -418,9 +749,34 @@ class BaseDistributionItem:
                 f"{self.item_label}: {message}",
                 exc_info=True
             )
-            return False
+            yield False
 
-    def _distribute(self):
+    def _set_progress_state(self, state: DistFileStates):
+        if not self._progress_dir:
+            return
+
+        progress_info = _read_progress_file(self._progress_dir)
+        if progress_info.get("progress_id") != self._progress_id:
+            # Ignore if wanted to set 'failed' state
+            if state == DistFileStates.failed:
+                return
+            raise DistributionProgressInterupted(
+                "Different process took over progress file."
+            )
+
+        progress_info["state"] = state.value
+        if state == DistFileStates.done:
+            # Update checksum if it was not set from server information
+            if not progress_info["checksum"] and self.checksum:
+                progress_info["checksum"] = self.checksum
+                progress_info["checksum_algorithm"] = self.checksum_algorithm
+            progress_info["dist_finished"] = (
+                datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+            )
+
+        _store_progress_file(self._progress_dir, progress_info)
+
+    def _distribute(self) -> Generator[bool, None, None]:
         if not self.sources:
             message = (
                 f"{self.item_label}: Don't have"
@@ -429,35 +785,106 @@ class BaseDistributionItem:
             self.log.error(message)
             self._error_msg = message
             self.state = UpdateState.MISS_SOURCE_FILES
+            yield True
             return
 
+        # Progress file
+        # - Check if other process/machine already started the job
+        if self._progress_dir:
+            result = False
+            for result in _wait_for_other_process(
+                self._progress_dir,
+                self._progress_id,
+                self.log
+            ):
+                if result is not None:
+                    # Give other distribution items option to start waiting
+                    #   cycle
+                    yield False
+                    break
+                yield False
+
+            if result is True:
+                self.state = UpdateState.UPDATED
+                self.log.info(f"{self.item_label}: Already distributed")
+                yield True
+                return
+
+            _create_progress_file(
+                self._progress_dir,
+                self._progress_id,
+                self.checksum,
+                self.checksum_algorithm,
+            )
+
+        # Download
         for source, source_progress in self.sources:
-            if self._process_source(source, source_progress):
+            yield False
+            result = False
+            for result in self._process_source(source, source_progress):
+                if result is not None:
+                    break
+
+            if result:
                 break
 
         last_progress = self._current_source_progress
         self._current_source_progress = None
         if self.state == UpdateState.UPDATED:
+            self._set_progress_state(DistFileStates.done)
             self._used_source_progress = last_progress
             self.log.info(f"{self.item_label}: Distributed")
-            return
-
-        self.log.error(f"{self.item_label}: Failed to distribute")
-        self._error_msg = "Failed to receive or install source files"
+        else:
+            self._set_progress_state(DistFileStates.failed)
+            self.log.error(f"{self.item_label}: Failed to distribute")
+            self._error_msg = "Failed to receive or install source files"
+        yield True
 
     def _post_distribute(self):
         pass
 
-    def distribute(self):
-        """Execute distribution logic."""
+    def is_distributed(self):
+        if not self.need_distribution:
+            return True
+        return self.state == UpdateState.UPDATED
 
-        if not self.need_distribution or self._dist_started:
+    def distribute(self) -> Generator[bool, None, None]:
+        """Execute distribution logic."""
+        if self.is_distributed() or self._dist_started:
+            yield True
             return
 
         self._dist_started = True
         try:
-            if self.state == UpdateState.OUTDATED:
-                self._distribute()
+            try:
+                for result in self._distribute():
+                    if result:
+                        break
+                    yield False
+
+            except DistributionProgressInterupted:
+                wait_result = False
+                for wait_result in _wait_for_other_process(
+                    self._progress_dir, self._progress_id, self.log
+                ):
+                    if wait_result is not None:
+                        break
+                    yield False
+
+                if wait_result:
+                    self.state = UpdateState.UPDATED
+                else:
+                    self.state = UpdateState.UPDATE_FAILED
+                    self._error_msg = (
+                        "Other process took over distribution and failed."
+                    )
+                    self._error_detail = (
+                        "Please try to start AYON again and contact"
+                        " administrator if issue persist."
+                    )
+                    self.log.error(
+                        f"{self.item_label}: {self._error_msg}"
+                    )
 
         except Exception as exc:
             self.state = UpdateState.UPDATE_FAILED
@@ -471,15 +898,18 @@ class BaseDistributionItem:
             )
 
         finally:
-            self._dist_finished = True
             if self.state == UpdateState.OUTDATED:
                 self.state = UpdateState.UPDATE_FAILED
                 self._error_msg = "Distribution failed"
 
             self._post_distribute()
+            yield True
 
 
-def create_tmp_file(suffix=None, prefix=None):
+def create_tmp_file(
+    suffix: Optional[str] = None,
+    prefix: Optional[str] = None
+) -> str:
     with tempfile.NamedTemporaryFile(
         suffix=suffix, prefix=prefix, delete=False
     ) as tmp:
@@ -490,7 +920,7 @@ def create_tmp_file(suffix=None, prefix=None):
 class InstallerDistributionItem(BaseDistributionItem):
     """Distribution of new version of AYON launcher/Installer."""
 
-    def __init__(self, cleanup_on_fail, *args, **kwargs):
+    def __init__(self, cleanup_on_fail: bool, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._cleanup_on_fail = cleanup_on_fail
         self._executable = None
@@ -498,40 +928,46 @@ class InstallerDistributionItem(BaseDistributionItem):
         self._installer_error = None
 
     @property
-    def executable(self):
+    def executable(self) -> Optional[str]:
         """Path to distributed ayon executable.
 
         Returns:
-            Union[str, None]: Path to executable path which was distributed.
-        """
+            Optional[str]: Path to executable path which was distributed.
 
+        """
         return self._executable
 
     @property
-    def installer_path(self):
+    def installer_path(self) -> Optional[str]:
         """Path to a distribution package/installer.
 
         This can be used as reference for user where to find downloaded
             installer on disk and distribute it manually.
 
         Returns:
-            Union[str, None]: Path to installer.
-        """
+            Optional[str]: Path to installer.
 
+        """
         return self._installer_path
 
     @property
-    def installer_error(self):
+    def installer_error(self) -> Optional[str]:
         """Known installer error that happened during distribution.
 
         Returns:
-            Union[str, None]: Message that will be shown to user and logged
+            Optional[str]: Message that will be shown to user and logged
                 out.
-        """
 
+        """
         return self._installer_error
 
-    def _find_windows_executable(self, log_output):
+    @property
+    def is_missing_permissions(self) -> bool:
+        return not _has_write_permissions(
+            os.path.dirname(os.path.dirname(sys.executable))
+        )
+
+    def _find_windows_executable(self, log_output: str):
         """Find executable path in log output.
 
         Setup exe should print out log output to a file where are described
@@ -543,8 +979,8 @@ class InstallerDistributionItem(BaseDistributionItem):
 
         Args:
             log_output (str): Output from installer log.
-        """
 
+        """
         exe_name = "ayon.exe"
         for line in log_output.splitlines():
             idx = line.find(exe_name)
@@ -577,41 +1013,13 @@ class InstallerDistributionItem(BaseDistributionItem):
             if os.path.exists(executable_path):
                 return executable_path
 
-    def _windows_root_require_permissions(self, dirpath):
-        while not os.path.exists(dirpath):
-            _dirpath = os.path.dirname(dirpath)
-            if _dirpath == dirpath:
-                print((
-                    "Failed to determine if root requires permissions."
-                    " The disk is probably not mounted."
-                ))
-                return False
-            dirpath = _dirpath
-        try:
-            # Attempt to create a temporary file in the folder
-            temp_file_path = os.path.join(dirpath, uuid.uuid4().hex)
-            with open(temp_file_path, "w"):
-                pass
-            os.remove(temp_file_path)  # Clean up temporary file
-            return False
-
-        except PermissionError:
-            return True
-
-        except BaseException as exc:
-            print((
-                "Failed to determine if root requires permissions."
-                "Unexpected error: {}"
-            ).format(exc))
-            return False
-
-    def _install_windows(self, filepath):
+    def _install_windows(self, filepath: str):
         """Install windows AYON launcher.
 
         Args:
-            filepath (str): Path to setup .exe file.
-        """
+            filepath (str): Path to setup.exe file.
 
+        """
         install_root = os.path.dirname(os.path.dirname(sys.executable))
 
         # A file where installer may store log output
@@ -620,10 +1028,7 @@ class InstallerDistributionItem(BaseDistributionItem):
         install_exe_tmp = create_tmp_file(suffix="ayon_install_dir")
         user_arg = "/CURRENTUSER"
         # Ask for admin permissions if user is not admin and
-        if (
-            not ctypes.windll.shell32.IsUserAnAdmin()
-            and self._windows_root_require_permissions(install_root)
-        ):
+        if not _has_write_permissions(install_root):
             if HEADLESS_MODE_ENABLED:
                 raise InstallerDistributionError((
                     "Installation requires administration permissions, which"
@@ -665,20 +1070,24 @@ class InstallerDistributionItem(BaseDistributionItem):
 
         self._executable = executable
 
-    def _install_linux(self, filepath):
+    def _install_linux(self, filepath: str):
         """Install linux AYON launcher.
 
-        Linux installations are just an archive file, so we attempt to unzip the new
-        installation one level up of the one being run.
+        Linux installations are just an archive file, so we attempt to unzip
+            the new installation one level up of the one being run.
 
         Args:
             filepath (str): Path to a .tar.gz file.
+
         """
         install_root = os.path.dirname(os.path.dirname(sys.executable))
 
-        self.log.info(f"Installing AYON launcher {filepath} into:\n{install_root}")
+        self.log.info(
+            f"Installing AYON launcher {filepath} into:\n{install_root}"
+        )
 
-        os.makedirs(install_root, exist_ok=True)
+        if not os.path.exists(install_root):
+            os.makedirs(install_root, exist_ok=True)
 
         try:
             extract_archive_file(filepath, install_root)
@@ -694,29 +1103,30 @@ class InstallerDistributionItem(BaseDistributionItem):
         self.log.info(f"Setting executable to {executable}")
         self._executable = executable
 
-    def _install_macos(self, filepath):
+    def _install_macos(self, filepath: str):
         """Install macOS AYON launcher.
 
         Args:
             filepath (str): Path to a .dmg file.
-        """
 
+        """
         import plistlib
 
         # Attach dmg file and read plist output (bytes)
         stdout = subprocess.check_output([
             "hdiutil", "attach", filepath, "-plist", "-nobrowse"
         ])
+        mounted_volumes = []
         try:
             # Parse plist output and find mounted volume
             attach_info = plistlib.loads(stdout)
-            mounted_volumes = []
             for entity in attach_info["system-entities"]:
                 mounted_volume = entity.get("mount-point")
                 if mounted_volume:
                     mounted_volumes.append(mounted_volume)
 
             # We do expect there is only one .app in .dmg file
+            src_path = None
             src_filename = None
             for mounted_volume in mounted_volumes:
                 for filename in os.listdir(mounted_volume):
@@ -725,14 +1135,16 @@ class InstallerDistributionItem(BaseDistributionItem):
                         src_path = os.path.join(mounted_volume, src_filename)
                         break
 
-            # Copy the .app file to /Applications
-            dst_dir = "/Applications"
-            dst_path = os.path.join(dst_dir, src_filename)
-            subprocess.run(["cp", "-rf", src_path, dst_dir])
+            if src_path is not None:
+                # Copy the .app file to /Applications
+                dst_dir = "/Applications"
+                dst_path = os.path.join(dst_dir, src_filename)
+                subprocess.run(["cp", "-rf", src_path, dst_dir])
 
         finally:
             # Detach mounted volume
-            subprocess.run(["hdiutil", "detach", mounted_volume])
+            for mounted_volume in mounted_volumes:
+                subprocess.run(["hdiutil", "detach", mounted_volume])
 
         # Find executable inside .app file and return its path
         contents_dir = os.path.join(dst_path, "Contents")
@@ -744,13 +1156,15 @@ class InstallerDistributionItem(BaseDistributionItem):
         else:
             parsed_plist = plistlib.readPlist(plist_filepath)
         executable_filename = parsed_plist.get("CFBundleExecutable")
-        return os.path.join(
+        self._executable = os.path.join(
             contents_dir, "MacOS", executable_filename
         )
 
     def _install_file(self, filepath):
         """Trigger installation installer file based on platform."""
 
+        # TODO consider using platform name in target folder
+        #  - for windows and linux
         platform_name = platform.system().lower()
         if platform_name == "windows":
             self._install_windows(filepath)
@@ -764,8 +1178,12 @@ class InstallerDistributionItem(BaseDistributionItem):
             )
 
     def _post_source_process(
-        self, filepath, source_data, source_progress, downloader
-    ):
+        self,
+        filepath: str,
+        source_data: dict[str, Any],
+        source_progress: DistributeTransferProgress,
+        downloader: SourceDownloader,
+    ) -> Generator[Optional[bool], None, None]:
         self._installer_path = filepath
         success = False
         try:
@@ -803,7 +1221,7 @@ class InstallerDistributionItem(BaseDistributionItem):
                 self.downloader_data
             )
 
-        return True
+        yield True
 
 
 class DistributionItem(BaseDistributionItem):
@@ -813,7 +1231,7 @@ class DistributionItem(BaseDistributionItem):
     unzip directory where the downloaded content is unzipped.
 
     Args:
-        unzip_dirpath (str): Path to directory where zip is downloaded.
+        target_dirpath (str): Path to directory where zip is downloaded.
         download_dirpath (str): Path to directory where file is unzipped.
         state (UpdateState): Initial state (UpdateState.UPDATED or
             UpdateState.OUTDATED).
@@ -823,31 +1241,61 @@ class DistributionItem(BaseDistributionItem):
             distribution item.
         downloader_data (Dict[str, Any]): More information for downloaders.
         item_label (str): Label used in log outputs (and in UI).
+        progress_dir (Optional[str]): Directory where progress is stored for
+            other processes.
         logger (logging.Logger): Logger object.
-    """
 
-    def __init__(self,unzip_dirpath, *args, **kwargs):
-        self.unzip_dirpath = unzip_dirpath
+    """
+    def __init__(
+        self, target_dirpath: str, unzip_temp_dir: str, *args, **kwargs
+    ):
+        self.target_dirpath = target_dirpath
+        self.unzip_temp_dir = unzip_temp_dir
         super().__init__(*args, **kwargs)
 
-    def _pre_source_process(self):
-        super()._pre_source_process()
-        unzip_dirpath = self.unzip_dirpath
+    @property
+    def is_missing_permissions(self) -> bool:
+        return not _has_write_permissions(self.target_dirpath)
 
-        # Remove directory if exists
-        if os.path.isdir(unzip_dirpath):
-            self.log.debug(f"Cleaning {unzip_dirpath}")
-            shutil.rmtree(unzip_dirpath)
+    def _post_source_process(
+        self,
+        filepath: str,
+        source_data: dict[str, Any],
+        source_progress: DistributeTransferProgress,
+        downloader: SourceDownloader,
+    ) -> Generator[Optional[bool], None, None]:
+        source_progress.set_unzip_started()
+        result = False
+        for result in _wait_for_other_process(
+            self._progress_dir, self._progress_id, self.log
+        ):
+            if result is not None:
+                break
+            yield None
+
+        if result is True:
+            self.state = UpdateState.UPDATED
+            yield True
+            return
+
+        filename = os.path.basename(self.target_dirpath)
+        unzip_dirpath = os.path.join(self.unzip_temp_dir, filename)
+        # NOTE This is a workaround for dependency packages
+        # TODO remove when dependency packages are not stored to directory
+        #   ending with .zip
+        if filepath == unzip_dirpath:
+            filedir = os.path.dirname(filepath)
+            ext = os.path.splitext(filepath)[-1]
+            new_filepath = os.path.join(filedir, f"{uuid.uuid4().hex}{ext}")
+            os.rename(filepath, new_filepath)
+            filepath = new_filepath
 
         # Create directory
         os.makedirs(unzip_dirpath, exist_ok=True)
+        change_permissions_recursive(unzip_dirpath)
 
-    def _post_source_process(
-        self, filepath, source_data, source_progress, downloader
-    ):
-        source_progress.set_unzip_started()
         try:
-            downloader.unzip(filepath, self.unzip_dirpath)
+            downloader.unzip(filepath, unzip_dirpath)
         except Exception:
             message = "Couldn't unzip source file"
             source_progress.set_failed(message)
@@ -855,24 +1303,146 @@ class DistributionItem(BaseDistributionItem):
                 f"{self.item_label}: {message}",
                 exc_info=True
             )
-            return False
+            yield False
+            return
+
         source_progress.set_unzip_finished()
 
-        return super()._post_source_process(
+        yield None
+
+        result = False
+        for result in _wait_for_other_process(
+            self._progress_dir, self._progress_id, self.log
+        ):
+            if result is not None:
+                break
+            yield None
+
+        if result is True:
+            self.state = UpdateState.UPDATED
+            yield True
+            return
+
+        failed = False
+        renamed_mapping = []
+
+        # Target directory can contain only distribution metadata file
+        #   anything else is temporarily moved to different directory
+        #   - next to target directory
+        # NOTE: We might validate if the content is exactly same?
+        tmp_subfolder = os.path.join(
+            os.path.dirname(self.target_dirpath), uuid.uuid4().hex
+        )
+        if not os.path.exists(self.target_dirpath):
+            os.makedirs(self.target_dirpath, exist_ok=True)
+            change_permissions_recursive(self.target_dirpath)
+
+        for name in os.listdir(self.target_dirpath):
+            if name == DIST_PROGRESS_FILENAME:
+                continue
+            if not os.path.exists(tmp_subfolder):
+                os.makedirs(tmp_subfolder, exist_ok=True)
+                change_permissions_recursive(tmp_subfolder)
+
+            current_path = os.path.join(self.target_dirpath, name)
+            new_path = os.path.join(tmp_subfolder, name)
+            try:
+                os.rename(current_path, new_path)
+            except Exception:
+                message = (
+                    "Target files already exist and can't be renamed."
+                )
+                source_progress.set_failed(message)
+                self.log.warning(
+                    f"{self.item_label}: {message}",
+                    exc_info=True
+                )
+                failed = True
+                break
+            renamed_mapping.append((current_path, new_path))
+
+        if failed:
+            # Rollback renamed files
+            for src_path, renamed_path in renamed_mapping:
+                os.rename(renamed_path, src_path)
+            yield False
+            return
+
+        yield None
+
+        # Move unzipped content to target directory
+        moved_paths = []
+        rename_attemps = 5
+        for subname in os.listdir(unzip_dirpath):
+            target_path = os.path.join(self.target_dirpath, subname)
+            src_path = os.path.join(unzip_dirpath, subname)
+            dst_path = os.path.join(self.target_dirpath, subname)
+            self.log.debug("Copying %s to %s", src_path, dst_path)
+            # Sometimes 'os.rename' crashes because of permissions issue
+            #   even though it can be done in a future attempt (probably some
+            #   service traversing through files blocking it?)
+            for attempt in range(rename_attemps):
+                try:
+                    os.rename(src_path, dst_path)
+                    break
+                except Exception:
+                    if attempt < (rename_attemps - 1):
+                        yield None
+                        continue
+                    message = (
+                        "Failed to move unzipped content to target directory."
+                    )
+                    source_progress.set_failed(message)
+                    self.log.warning(
+                        f"{self.item_label}: {message}",
+                        exc_info=True
+                    )
+                    failed = True
+
+            if failed:
+                break
+
+            moved_paths.append(target_path)
+
+        # Handle failed movement
+        if failed:
+            # Remove moved files
+            for path in moved_paths:
+                if os.path.isfile(path):
+                    os.remove(path)
+                else:
+                    shutil.rmtree(path)
+
+            # Rename renamed files back to original
+            for src_path, renamed_path in renamed_mapping:
+                os.rename(renamed_path, src_path)
+
+        # Remove temp subfolder used for renaming
+        if os.path.exists(tmp_subfolder):
+            shutil.rmtree(tmp_subfolder)
+
+        if failed:
+            yield False
+            return
+
+        yield from super()._post_source_process(
             filepath, source_data, source_progress, downloader
         )
 
     def _post_distribute(self):
         if (
             self.state != UpdateState.UPDATED
-            and self.unzip_dirpath
-            and os.path.isdir(self.unzip_dirpath)
+            and self.target_dirpath
+            and os.path.isdir(self.target_dirpath)
         ):
-            self.log.debug(f"Cleaning {self.unzip_dirpath}")
-            shutil.rmtree(self.unzip_dirpath)
+            self.log.debug(f"Cleaning {self.target_dirpath}")
+            # TODO use '_clean_dist_dir' when backwards compatibility with
+            #   previous versions of AYON launchers is not needed.
+            # _clean_dist_dir(self.target_dirpath)
+            shutil.rmtree(self.target_dirpath)
 
 
-class AyonDistribution:
+class AYONDistribution:
     """Distribution control.
 
     Receive information from server what addons and dependency packages
@@ -885,11 +1455,13 @@ class AyonDistribution:
         dependency_dirpath (Optional[str]): Where dependencies will be stored.
         dist_factory (Optional[DownloadFactory]): Factory which cares about
             downloading of items based on source type.
-        addons_info (Optional[list[dict[str, Any]]): List of prepared
+        installers_info (Optional[list[dict[str, Any]]]): List of prepared
+            installers' info.
+        addons_info (Optional[list[dict[str, Any]]]): List of prepared
             addons' info.
-        dependency_packages_info (Optional[list[dict[str, Any]]): Info
+        dependency_packages_info (Optional[list[dict[str, Any]]]): Info
             about packages from server.
-        bundles_info (Optional[Dict[str, Any]]): Info about
+        bundles_info (Optional[dict[str, Any]]): Info about
             bundles.
         bundle_name (Optional[str]): Name of bundle to use. If not passed
             an environment variable 'AYON_BUNDLE_NAME' is checked for value.
@@ -899,34 +1471,44 @@ class AyonDistribution:
             If not passed, 'is_staging_enabled' is used as default value.
         use_dev (Optional[bool]): Use develop versions of an addon.
             If not passed, 'is_dev_mode_enabled' is used as default value.
-        skip_installer_dist (Optional[bool]): Skip installer distribution. This
+        skip_installer_dist (bool): Skip installer distribution. This
             is for testing purposes and for running from code.
-    """
 
+    """
     def __init__(
         self,
-        addon_dirpath=None,
-        dependency_dirpath=None,
-        dist_factory=None,
-        installers_info=NOT_SET,
-        addons_info=NOT_SET,
-        dependency_packages_info=NOT_SET,
-        bundles_info=NOT_SET,
-        bundle_name=NOT_SET,
-        use_staging=None,
-        use_dev=None,
-        active_user=None,
-        skip_installer_dist=False,
+        addon_dirpath: Optional[str] = None,
+        dependency_dirpath: Optional[str] = None,
+        dist_factory: Optional[DownloadFactory] = None,
+        installers_info: Optional[list[dict[str, Any]]] = NOT_SET,
+        addons_info: Optional[list[dict[str, Any]]] = NOT_SET,
+        dependency_packages_info: Optional[list[dict[str, Any]]] = NOT_SET,
+        bundles_info: Optional[list[dict[str, Any]]] = NOT_SET,
+        bundle_name: Optional[str] = NOT_SET,
+        use_staging: Optional[bool] = None,
+        use_dev: Optional[bool] = None,
+        active_user: Optional[bool] = None,
+        skip_installer_dist: bool = False,
     ):
         self._log = None
 
         self._dist_started = False
-        self._dist_finished = False
 
         self._addons_dirpath = addon_dirpath or get_addons_dir()
         self._dependency_dirpath = dependency_dirpath or get_dependencies_dir()
         self._dist_factory = (
             dist_factory or get_default_download_factory()
+        )
+
+        # Where addon zip files and dependency packages are downloaded
+        self._dist_download_dirs = []
+        self._dist_unzip_dirs = []
+
+        self._dist_addons_unzip_temp = os.path.join(
+            self._addons_dirpath, ".unzip_temp"
+        )
+        self._dist_dep_packages_unzip_temp = os.path.join(
+            self._dependency_dirpath, ".unzip_temp"
         )
 
         if bundle_name is NOT_SET:
@@ -980,14 +1562,14 @@ class AyonDistribution:
         self._bundle = NOT_SET
 
     @property
-    def active_user(self):
+    def active_user(self) -> str:
         if self._active_user is None:
             user = ayon_api.get_user()
             self._active_user = user["name"]
         return self._active_user
 
     @property
-    def use_staging(self):
+    def use_staging(self) -> bool:
         """Staging version of a bundle should be used.
 
         This value is completely ignored if specific bundle name should
@@ -995,8 +1577,8 @@ class AyonDistribution:
 
         Returns:
             bool: True if staging version should be used.
-        """
 
+        """
         if self._use_staging is None:
             self._use_staging = is_staging_enabled()
 
@@ -1005,7 +1587,7 @@ class AyonDistribution:
         return self._use_staging
 
     @property
-    def use_dev(self):
+    def use_dev(self) -> bool:
         """Develop version of a bundle should be used.
 
         This value is completely ignored if specific bundle name should
@@ -1013,8 +1595,8 @@ class AyonDistribution:
 
         Returns:
             bool: True if staging version should be used.
-        """
 
+        """
         if self._use_dev is None:
             if self._bundle_name is NOT_SET:
                 self._use_dev = is_dev_mode_enabled()
@@ -1034,95 +1616,96 @@ class AyonDistribution:
         return self._use_dev
 
     @property
-    def log(self):
+    def log(self) -> logging.Logger:
         """Helper to access logger.
 
         Returns:
              logging.Logger: Logger instance.
+
         """
         if self._log is None:
             self._log = logging.getLogger(self.__class__.__name__)
         return self._log
 
     @property
-    def bundles_info(self):
+    def bundles_info(self) -> list[dict[str, Any]]:
         """
 
         Returns:
-            dict[str, dict[str, Any]]: Bundles information from server.
-        """
+            list[dict[str, Any]]: Bundles information from server.
 
+        """
         if self._bundles_info is NOT_SET:
-            self._bundles_info = ayon_api.get_bundles()
+            self._bundles_info = ayon_api.get_bundles()["bundles"]
         return self._bundles_info
 
     @property
-    def bundle_items(self):
+    def bundle_items(self) -> list[Bundle]:
         """
 
         Returns:
             list[Bundle]: List of bundles info.
-        """
 
+        """
         if self._bundle_items is NOT_SET:
             self._bundle_items = [
                 Bundle.from_dict(info)
-                for info in self.bundles_info["bundles"]
+                for info in self.bundles_info
             ]
         return self._bundle_items
 
     @property
-    def production_bundle(self):
+    def production_bundle(self) -> Optional[Bundle]:
         """
 
         Returns:
-            Union[Bundle, None]: Bundle that should be used in production.
-        """
+            Optional[Bundle]: Bundle that should be used in production.
 
+        """
         if self._production_bundle is NOT_SET:
             self._prepare_bundles()
         return self._production_bundle
 
     @property
-    def staging_bundle(self):
+    def staging_bundle(self) -> Optional[Bundle]:
         """
 
         Returns:
-            Union[Bundle, None]: Bundle that should be used in staging.
-        """
+            Optional[Bundle]: Bundle that should be used in staging.
 
+        """
         if self._staging_bundle is NOT_SET:
             self._prepare_bundles()
         return self._staging_bundle
 
     @property
-    def dev_bundle(self):
+    def dev_bundle(self) -> Optional[Bundle]:
         """
 
         Returns:
-            Union[Bundle, None]: Bundle that should be used in dev.
-        """
+            Optional[Bundle]: Bundle that should be used in dev.
 
+        """
         if self._dev_bundle is NOT_SET:
             self._prepare_bundles()
         return self._dev_bundle
 
     @property
-    def bundle_to_use(self):
+    def bundle_to_use(self) -> Optional[Bundle]:
         """Bundle that will be used for distribution.
 
         Bundle that should be used can be affected by 'bundle_name'
             or 'use_staging'.
 
         Returns:
-            Union[Bundle, None]: Bundle that will be used for distribution
+            Optional[Bundle]: Bundle that will be used for distribution
                 or None.
 
         Raises:
             BundleNotFoundError: When bundle name to use is defined
                 but is not available on server.
-        """
 
+        """
         if self._bundle is not NOT_SET:
             return self._bundle
 
@@ -1152,37 +1735,37 @@ class AyonDistribution:
         return self._bundle
 
     @property
-    def bundle_name_to_use(self):
+    def bundle_name_to_use(self) -> Optional[str]:
         """Name of bundle that will be used for distribution.
 
         Returns:
-            Union[str, None]: Name of bundle that will be used for
+            Optional[str]: Name of bundle that will be used for
                 distribution.
-        """
 
+        """
         bundle = self.bundle_to_use
         return None if bundle is None else bundle.name
 
     @property
-    def installers_info(self):
+    def installers_info(self) -> list[dict[str, Any]]:
         """Installers information from server.
 
         Returns:
             list[dict[str, Any]]: Installers information from server.
-        """
 
+        """
         if self._installers_info is NOT_SET:
             self._installers_info = ayon_api.get_installers()["installers"]
         return self._installers_info
 
     @property
-    def installer_items(self):
+    def installer_items(self) -> list[Installer]:
         """Installers as objects.
 
         Returns:
             list[Installer]: List of installers info from server.
-        """
 
+        """
         if self._installer_items is NOT_SET:
             self._installer_items = [
                 Installer.from_dict(info)
@@ -1191,14 +1774,14 @@ class AyonDistribution:
         return self._installer_items
 
     @property
-    def expected_installer_version(self):
+    def expected_installer_version(self) -> Optional[str]:
         """Excepted installer version.
 
         Returns:
-            Union[str, None]: Expected installer version or None defined by
+            Optional[str]: Expected installer version or None defined by
                 bundle that should be used.
-        """
 
+        """
         if self._expected_installer_version is not NOT_SET:
             return self._expected_installer_version
 
@@ -1208,7 +1791,7 @@ class AyonDistribution:
         return version
 
     @property
-    def need_installer_change(self):
+    def need_installer_change(self) -> bool:
         """Installer should be changed.
 
         Current installer is using different version than what is expected
@@ -1216,8 +1799,8 @@ class AyonDistribution:
 
         Returns:
             bool: True if installer should be changed.
-        """
 
+        """
         if self._skip_installer_dist:
             return False
 
@@ -1225,7 +1808,7 @@ class AyonDistribution:
         return version != self.expected_installer_version
 
     @property
-    def need_installer_distribution(self):
+    def need_installer_distribution(self) -> bool:
         """Installer distribution is needed.
 
         Todos:
@@ -1233,39 +1816,39 @@ class AyonDistribution:
 
         Returns:
             bool: True if installer distribution is needed.
-        """
 
+        """
         if not self.need_installer_change:
             return False
 
         return self.installer_executable is None
 
     @property
-    def installer_dist_error(self):
+    def installer_dist_error(self) -> Optional[str]:
         """Installer distribution error message.
 
         Returns:
-              Union[str, None]: Error that happened during installer
+              Optional[str]: Error that happened during installer
                 distribution.
-        """
 
+        """
         return self._installer_dist_error
 
     @property
-    def installer_filepath(self):
+    def installer_filepath(self) -> Optional[str]:
         """Path to a distribution package/installer.
 
         This can be used as reference for user where to find downloaded
             installer on disk and distribute it manually.
 
         Returns:
-            Union[str, None]: Path to installer.
-        """
+            Optional[str]: Path to installer.
 
+        """
         return self._installer_filepath
 
     @property
-    def installer_executable(self):
+    def installer_executable(self) -> Optional[str]:
         """Path to installer executable that should be used.
 
         Notes:
@@ -1273,11 +1856,11 @@ class AyonDistribution:
                 called 'ayon_executable'?
 
         Returns:
-            Union[str, None]: Path to installer executable that should be
+            Optional[str]: Path to installer executable that should be
                 used. None if executable is not found and must be distributed
                 or bundle does not have defined an installer to use.
-        """
 
+        """
         if self._installer_executable is not NOT_SET:
             return self._installer_executable
 
@@ -1293,6 +1876,7 @@ class AyonDistribution:
             if filename == "ayon_console.exe":
                 current_executable = os.path.join(root, "ayon.exe")
 
+            # TODO look to expected target install directory too
             executables_info = get_executables_info_by_version(
                 self.expected_installer_version)
             for executable_info in executables_info:
@@ -1318,8 +1902,8 @@ class AyonDistribution:
 
         Returns:
             Union[Installer, None]: Installer information item.
-        """
 
+        """
         if self._installer_item is not NOT_SET:
             return self._installer_item
 
@@ -1364,6 +1948,7 @@ class AyonDistribution:
         if not downloads_dir or not os.path.exists(downloads_dir):
             tmp_used = True
             downloads_dir = tempfile.mkdtemp(prefix="ayon_installer")
+            change_permissions_recursive(downloads_dir)
 
         dist_item = None
         try:
@@ -1376,9 +1961,25 @@ class AyonDistribution:
                 self._dist_factory,
                 list(installer_item.sources),
                 downloader_data,
-                f"Installer {installer_item.version}"
+                f"Installer {installer_item.version}",
             )
-            dist_item.distribute()
+
+            if (
+                platform.system().lower() != "windows"
+                and dist_item.is_missing_permissions
+            ):
+                self._installer_dist_error = (
+                    "Your user does not have required permissions to update"
+                    " AYON launcher."
+                    " Please contact your administrator, or use user"
+                    " with permissions."
+                )
+                return
+
+            for result in dist_item.distribute():
+                if result:
+                    break
+
             self._installer_executable = dist_item.executable
             if dist_item.installer_error is not None:
                 self._installer_dist_error = dist_item.installer_error
@@ -1413,13 +2014,13 @@ class AyonDistribution:
                 shutil.rmtree(downloads_dir)
 
     @property
-    def addons_info(self):
+    def addons_info(self) -> dict[str, dict[str, Any]]:
         """Server information about available addons.
 
         Returns:
-            Dict[str, dict[str, Any]: Addon info by addon name.
-        """
+            dict[str, dict[str, Any]]: Addon info by addon name.
 
+        """
         if self._addons_info is NOT_SET:
             # Use details to get information about client.zip
             server_info = ayon_api.get_addons_info(details=True)
@@ -1427,16 +2028,16 @@ class AyonDistribution:
         return self._addons_info
 
     @property
-    def addon_items(self):
+    def addon_items(self) -> dict[str, AddonInfo]:
         """Information about available addons on server.
 
         Addons may require distribution of files. For those addons will be
         created 'DistributionItem' handling distribution itself.
 
         Returns:
-            Dict[str, AddonInfo]: Addon info object by addon name.
-        """
+            dict[str, AddonInfo]: Addon info object by addon name.
 
+        """
         if self._addon_items is NOT_SET:
             addons_info = {}
             for addon in self.addons_info:
@@ -1446,7 +2047,7 @@ class AyonDistribution:
         return self._addon_items
 
     @property
-    def dependency_packages_info(self):
+    def dependency_packages_info(self) -> list[dict[str, Any]]:
         """Server information about available dependency packages.
 
         Notes:
@@ -1455,21 +2056,21 @@ class AyonDistribution:
 
         Returns:
             list[dict[str, Any]]: Dependency packages information.
-        """
 
+        """
         if self._dependency_packages_info is NOT_SET:
             self._dependency_packages_info = (
                 ayon_api.get_dependency_packages())["packages"]
         return self._dependency_packages_info
 
     @property
-    def dependency_packages_items(self):
+    def dependency_packages_items(self) -> dict[str, DependencyItem]:
         """Dependency packages as objects.
 
         Returns:
             dict[str, DependencyItem]: Dependency packages as objects by name.
-        """
 
+        """
         if self._dependency_packages_items is NOT_SET:
             dependenc_package_items = {}
             for item in self.dependency_packages_info:
@@ -1479,14 +2080,14 @@ class AyonDistribution:
         return self._dependency_packages_items
 
     @property
-    def dependency_package_item(self):
+    def dependency_package_item(self) -> Optional[DependencyItem]:
         """Dependency package item that should be used by bundle.
 
         Returns:
-            Union[None, Dict[str, Any]]: None if bundle does not have
+            Optional[DependencyItem]: None if bundle does not have
                 specified dependency package.
-        """
 
+        """
         if self._dependency_package_item is NOT_SET:
             dependency_package_item = None
             bundle = self.bundle_to_use
@@ -1514,7 +2115,7 @@ class AyonDistribution:
         self._staging_bundle = staging_bundle
         self._dev_bundle = dev_bundle
 
-    def _prepare_current_addon_dist_items(self):
+    def _prepare_current_addon_dist_items(self) -> list[dict[str, Any]]:
         addons_metadata = self.get_addons_metadata()
         output = []
         addon_versions = {}
@@ -1524,11 +2125,16 @@ class AyonDistribution:
             dev_addons = bundle.addons_dev_info
             addon_versions = bundle.addon_versions
 
+        unzip_temp = os.path.join(self._addons_dirpath, ".unzip_temp")
+        if not os.path.exists(unzip_temp):
+            os.makedirs(unzip_temp)
+            change_permissions_recursive(unzip_temp)
+
         for addon_name, addon_item in self.addon_items.items():
             # Dev mode can redirect addon directory elsewhere
             if self.use_dev:
-                dev_addon_info = dev_addons.get(addon_name) or {}
-                if dev_addon_info.get("enabled") is True:
+                dev_addon_info = dev_addons.get(addon_name)
+                if dev_addon_info is not None and dev_addon_info.enabled:
                     continue
 
             addon_version = addon_versions.get(addon_name)
@@ -1551,28 +2157,44 @@ class AyonDistribution:
             full_name = addon_version_item.full_name
             addon_dest = os.path.join(self._addons_dirpath, full_name)
             self.log.debug(f"Checking {full_name} in {addon_dest}")
-            addon_in_metadata = (
-                addon_name in addons_metadata
-                and addon_version_item.version in addons_metadata[addon_name]
-            )
-            if addon_in_metadata and os.path.isdir(addon_dest):
-                self.log.debug(
-                    f"Addon version folder {addon_dest} already exists."
-                )
-                state = UpdateState.UPDATED
-
+            progress_info = _read_progress_file(addon_dest)
+            state = UpdateState.OUTDATED
+            if progress_info:
+                if progress_info.get("state") == DistFileStates.done.value:
+                    state = UpdateState.UPDATED
             else:
-                state = UpdateState.OUTDATED
+                addon_in_metadata = (
+                    addon_name in addons_metadata
+                    and addon_version_item.version in (
+                        addons_metadata[addon_name]
+                    )
+                )
+                if addon_in_metadata and os.path.isdir(addon_dest):
+                    self.log.debug(
+                        f"Addon version folder {addon_dest} already exists."
+                    )
+                    state = UpdateState.UPDATED
+                    # Auto-create addons dist file extracted with old versions
+                    _create_progress_file(
+                        addon_dest,
+                        uuid.uuid4().hex,
+                        addon_version_item.checksum,
+                        addon_version_item.checksum_algorithm,
+                    )
 
+            download_dir = _get_dist_download_dir(uuid.uuid4().hex)
+            unzip_dir = os.path.join(unzip_temp, uuid.uuid4().hex)
+            self._dist_download_dirs.append(download_dir)
+            self._dist_unzip_dirs.append(unzip_dir)
             downloader_data = {
                 "type": "addon",
                 "name": addon_name,
                 "version": addon_version
             }
-
             dist_item = DistributionItem(
                 addon_dest,
-                download_dirpath=addon_dest,
+                unzip_dir,
+                download_dirpath=download_dir,
                 state=state,
                 checksum=addon_version_item.checksum,
                 checksum_algorithm=addon_version_item.checksum_algorithm,
@@ -1580,6 +2202,7 @@ class AyonDistribution:
                 sources=list(addon_version_item.sources),
                 downloader_data=downloader_data,
                 item_label=full_name,
+                progress_dir=addon_dest,
                 logger=self.log
             )
             output.append({
@@ -1591,7 +2214,7 @@ class AyonDistribution:
             })
         return output
 
-    def _prepare_dependency_progress(self):
+    def _prepare_dependency_progress(self) -> Optional[DistributionItem]:
         package = self.dependency_package_item
         if package is None:
             return None
@@ -1602,19 +2225,48 @@ class AyonDistribution:
             "name": package.filename,
             "platform": package.platform_name
         }
-        zip_dir = package_dir = os.path.join(
+        package_dir = os.path.join(
             self._dependency_dirpath, package.filename
         )
-        self.log.debug(f"Checking {package.filename} in {package_dir}")
+        # Future compatibility for dependency packages without .zip in dirname
+        new_basename = os.path.splitext(package.filename)[0]
+        new_package_dir = os.path.join(self._dependency_dirpath, new_basename)
 
-        if not os.path.isdir(package_dir) or package.filename not in metadata:
-            state = UpdateState.OUTDATED
-        else:
+        download_dir = _get_dist_download_dir(uuid.uuid4().hex)
+        unzip_dir = os.path.join(self._dependency_dirpath, uuid.uuid4().hex)
+        self._dist_download_dirs.append(download_dir)
+        self._dist_unzip_dirs.append(unzip_dir)
+
+        self.log.debug(f"Checking {package.filename} in {package_dir}")
+        state = UpdateState.OUTDATED
+        progress_info = _read_progress_file(package_dir)
+        new_progress_info = _read_progress_file(new_package_dir)
+        if progress_info:
+            if progress_info.get("state") == DistFileStates.done.value:
+                state = UpdateState.UPDATED
+
+        elif new_progress_info:
+            if new_progress_info.get("state") == DistFileStates.done.value:
+                state = UpdateState.UPDATED
+                package_dir = new_package_dir
+
+        elif (
+            os.path.isdir(package_dir)
+            and package.filename in metadata
+        ):
             state = UpdateState.UPDATED
+            # Autofix dependency packages extracted with old versions
+            _create_progress_file(
+                package_dir,
+                uuid.uuid4().hex,
+                package.checksum,
+                package.checksum_algorithm,
+            )
 
         return DistributionItem(
-            zip_dir,
-            download_dirpath=package_dir,
+            package_dir,
+            unzip_dir,
+            download_dirpath=download_dir,
             state=state,
             checksum=package.checksum,
             checksum_algorithm=package.checksum_algorithm,
@@ -1622,10 +2274,11 @@ class AyonDistribution:
             sources=package.sources,
             downloader_data=downloader_data,
             item_label=os.path.splitext(package.filename)[0],
+            progress_dir=package_dir,
             logger=self.log,
         )
 
-    def get_addon_dist_items(self):
+    def get_addon_dist_items(self) -> list[dict[str, Any]]:
         """Addon distribution items.
 
         These items describe source files required by addon to be available on
@@ -1647,14 +2300,14 @@ class AyonDistribution:
 
         Returns:
              list[dict[str, Any]]: Distribution items with addon version item.
-        """
 
+        """
         if self._addon_dist_items is NOT_SET:
             self._addon_dist_items = (
                 self._prepare_current_addon_dist_items())
         return self._addon_dist_items
 
-    def get_dependency_dist_item(self):
+    def get_dependency_dist_item(self) -> Optional[DistributionItem]:
         """Dependency package distribution item.
 
         Item describe source files required by server to be available on
@@ -1665,15 +2318,15 @@ class AyonDistribution:
         package.
 
         Returns:
-            Union[None, DistributionItem]: Dependency item or None if server
+            Optional[DistributionItem]: Dependency item or None if server
                 does not have specified any dependency package.
-        """
 
+        """
         if self._dependency_dist_item is NOT_SET:
             self._dependency_dist_item = self._prepare_dependency_progress()
         return self._dependency_dist_item
 
-    def get_dependency_metadata_filepath(self):
+    def get_dependency_metadata_filepath(self) -> str:
         """Path to distribution metadata file.
 
         Metadata contain information about distributed packages, used source,
@@ -1681,11 +2334,11 @@ class AyonDistribution:
 
         Returns:
             str: Path to a file where dependency package metadata are stored.
-        """
 
+        """
         return os.path.join(self._dependency_dirpath, "dependency.json")
 
-    def get_addons_metadata_filepath(self):
+    def get_addons_metadata_filepath(self) -> str:
         """Path to addons metadata file.
 
         Metadata contain information about distributed addons, used sources,
@@ -1693,11 +2346,15 @@ class AyonDistribution:
 
         Returns:
             str: Path to a file where addons metadata are stored.
-        """
 
+        """
         return os.path.join(self._addons_dirpath, "addons.json")
 
-    def read_metadata_file(self, filepath, default_value=None):
+    @staticmethod
+    def read_metadata_file(
+        filepath: str,
+        default_value: Optional[Any] = None,
+    ) -> dict[str, Any]:
         """Read json file from path.
 
         Method creates the file when does not exist with default value.
@@ -1724,35 +2381,40 @@ class AyonDistribution:
             data = default_value
         return data
 
-    def save_metadata_file(self, filepath, data):
+    @staticmethod
+    def save_metadata_file(filepath: str, data: dict[str, Any]):
         """Store data to json file.
 
         Method creates the file when does not exist.
 
         Args:
             filepath (str): Path to json file.
-            data (Union[Dict[str, Any], List[Any]]): Data to store into file.
+            data (dict[str, Any]): Data to store into file.
 
         """
-        os.makedirs(os.path.dirname(filepath), exist_ok=True)
+        dirpath = os.path.dirname(filepath)
+        os.makedirs(dirpath, exist_ok=True)
         with open(filepath, "w") as stream:
             json.dump(data, stream, indent=4)
+        change_permissions_recursive(dirpath)
 
-    def get_dependency_metadata(self):
+    def get_dependency_metadata(self) -> dict[str, Any]:
         filepath = self.get_dependency_metadata_filepath()
         return self.read_metadata_file(filepath, {})
 
-    def update_dependency_metadata(self, package_name, data):
+    def update_dependency_metadata(
+        self, package_name: str, data: dict[str, Any]
+    ):
         dependency_metadata = self.get_dependency_metadata()
         dependency_metadata[package_name] = data
         filepath = self.get_dependency_metadata_filepath()
         self.save_metadata_file(filepath, dependency_metadata)
 
-    def get_addons_metadata(self):
+    def get_addons_metadata(self) -> dict[str, Any]:
         filepath = self.get_addons_metadata_filepath()
         return self.read_metadata_file(filepath, {})
 
-    def update_addons_metadata(self, addons_information):
+    def update_addons_metadata(self, addons_information: dict[str, Any]):
         if not addons_information:
             return
         addons_metadata = self.get_addons_metadata()
@@ -1767,9 +2429,20 @@ class AyonDistribution:
 
     def finish_distribution(self):
         """Store metadata about distributed items."""
+        for tmp_dir in (
+            self._dist_download_dirs + self._dist_unzip_dirs
+        ):
+            if os.path.exists(tmp_dir):
+                try:
+                    shutil.rmtree(tmp_dir)
+                except Exception:
+                    pass
 
-        self._dist_finished = True
         stored_time = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        # TODO store dependencies info inside dependencies folder instead
+        #   of having one file
+        # - the file can be used to track progress and find out if other
+        #   process is already working on distribution
         dependency_dist_item = self.get_dependency_dist_item()
         if (
             dependency_dist_item is not None
@@ -1788,6 +2461,10 @@ class AyonDistribution:
                 }
                 self.update_dependency_metadata(package.filename, data)
 
+        # TODO store addon info inside addon folder instead of having one
+        #   of having one file
+        # - the file can be used to track progress and find out if other
+        #   process is already working on distribution
         addons_info = {}
         for item in self.get_addon_dist_items():
             dist_item = item["dist_item"]
@@ -1813,7 +2490,11 @@ class AyonDistribution:
 
         self.update_addons_metadata(addons_info)
 
-    def get_all_distribution_items(self):
+        _cleanup_dist_download_dirs()
+        _cleanup_dist_expire_dirs(self._dist_addons_unzip_temp)
+        _cleanup_dist_expire_dirs(self._dist_dep_packages_unzip_temp)
+
+    def get_all_distribution_items(self) -> list[DistributionItem]:
         """Distribution items required by server.
 
         Items contain dependency package item and all addons that are enabled
@@ -1822,9 +2503,9 @@ class AyonDistribution:
         Items can be already available on machine.
 
         Returns:
-            List[DistributionItem]: Distribution items required by server.
-        """
+            list[DistributionItem]: Distribution items required by server.
 
+        """
         output = [
             item["dist_item"]
             for item in self.get_addon_dist_items()
@@ -1836,13 +2517,13 @@ class AyonDistribution:
         return output
 
     @property
-    def need_distribution(self):
+    def need_distribution(self) -> bool:
         """Distribution is needed.
 
         Returns:
             bool: True if any distribution is needed.
-        """
 
+        """
         if self.need_installer_change:
             if self.need_installer_distribution:
                 return True
@@ -1853,7 +2534,16 @@ class AyonDistribution:
                 return True
         return False
 
-    def distribute(self, threaded=False):
+    @property
+    def is_missing_permissions(self):
+        # Do not validate installer (launcher) distribution as that is
+        #   reported with '_installer_dist_error'
+        for item in self.get_all_distribution_items():
+            if item.need_distribution and item.is_missing_permissions:
+                return True
+        return False
+
+    def distribute(self):
         """Distribute all missing items.
 
         Method will try to distribute all items that are required by server.
@@ -1861,10 +2551,7 @@ class AyonDistribution:
         This method does not handle failed items. To validate the result call
         'validate_distribution' when this method finishes.
 
-        Args:
-            threaded (bool): Distribute items in threads.
         """
-
         if self._dist_started:
             raise RuntimeError("Distribution already started")
         self._dist_started = True
@@ -1874,29 +2561,41 @@ class AyonDistribution:
                 self.distribute_installer()
             return
 
-        threads = collections.deque()
-        for item in self.get_all_distribution_items():
-            if threaded:
-                threads.append(threading.Thread(target=item.distribute))
-            else:
-                item.distribute()
+        dist_items = []
+        for dist_item in self.get_all_distribution_items():
+            if not dist_item.is_distributed():
+                dist_items.append(dist_item)
+                _create_dist_expire_file(dist_item.download_dirpath)
 
-        while threads:
-            thread = threads.popleft()
-            if thread.is_alive():
-                threads.append(thread)
-            else:
-                thread.join()
+        running_items = collections.deque()
+        for item in dist_items:
+            running_items.append(item.distribute())
 
-        self.finish_distribution()
+        if running_items:
+            running_items.append(None)
+
+        try:
+            while running_items:
+                running_item = running_items.popleft()
+                if running_item is None:
+                    if running_items:
+                        running_items.append(None)
+                        time.sleep(0.02)
+                    continue
+
+                if not next(running_item):
+                    running_items.append(running_item)
+
+        finally:
+            self.finish_distribution()
 
     def validate_distribution(self):
         """Check if all required distribution items are distributed.
 
         Raises:
             RuntimeError: Any of items is not available.
-        """
 
+        """
         invalid = []
         dependency_package = self.get_dependency_dist_item()
         if (
@@ -1917,7 +2616,7 @@ class AyonDistribution:
             ", ".join([f'"{item}"' for item in invalid])
         ))
 
-    def get_sys_paths(self):
+    def get_sys_paths(self) -> list[str]:
         """Get all paths to python packages that should be added to path.
 
         These packages will be added only to 'sys.path' and not into
@@ -1929,22 +2628,22 @@ class AyonDistribution:
                 dependencies (OpenTimelineIO, Pillow, etc.).
 
         Returns:
-            List[str]: Paths that should be added to 'sys.path'.
-        """
+            list[str]: Paths that should be added to 'sys.path'.
 
+        """
         output = []
         dependency_dist_item = self.get_dependency_dist_item()
         if dependency_dist_item is not None:
             runtime_dir = None
-            unzip_dirpath = dependency_dist_item.unzip_dirpath
-            if unzip_dirpath:
-                runtime_dir = os.path.join(unzip_dirpath, "runtime")
+            target_dirpath = dependency_dist_item.target_dirpath
+            if target_dirpath:
+                runtime_dir = os.path.join(target_dirpath, "runtime")
 
             if runtime_dir and os.path.exists(runtime_dir):
                 output.append(runtime_dir)
         return output
 
-    def get_python_paths(self):
+    def get_python_paths(self) -> list[str]:
         """Get all paths to python packages that should be added to python.
 
         These paths lead to addon directories and python dependencies in
@@ -1953,31 +2652,31 @@ class AyonDistribution:
         Returns:
             List[str]: Paths that should be added to 'sys.path' and
                 'PYTHONPATH'.
-        """
 
+        """
         output = []
         for item in self.get_addon_dist_items():
             dist_item = item["dist_item"]
             if dist_item.state != UpdateState.UPDATED:
                 continue
-            unzip_dirpath = dist_item.unzip_dirpath
-            if unzip_dirpath and os.path.exists(unzip_dirpath):
-                output.append(unzip_dirpath)
+            target_dirpath = dist_item.target_dirpath
+            if target_dirpath and os.path.exists(target_dirpath):
+                output.append(target_dirpath)
 
         output.extend(self._get_dev_sys_paths())
 
         dependency_dist_item = self.get_dependency_dist_item()
         if dependency_dist_item is not None:
             dependencies_dir = None
-            unzip_dirpath = dependency_dist_item.unzip_dirpath
-            if unzip_dirpath:
-                dependencies_dir = os.path.join(unzip_dirpath, "dependencies")
+            target_dirpath = dependency_dist_item.target_dirpath
+            if target_dirpath:
+                dependencies_dir = os.path.join(target_dirpath, "dependencies")
 
             if dependencies_dir and os.path.exists(dependencies_dir):
                 output.append(dependencies_dir)
         return output
 
-    def _get_dev_sys_paths(self):
+    def _get_dev_sys_paths(self) -> list[str]:
         output = []
         if not self.use_dev:
             return output
@@ -1995,9 +2694,9 @@ class AyonDistribution:
             if addon_version is None:
                 continue
 
-            dev_addon_info = dev_addons.get(addon_name) or {}
-            if dev_addon_info.get("enabled") is True:
-                output.append(dev_addon_info["path"])
+            dev_addon_info = dev_addons.get(addon_name)
+            if dev_addon_info is not None and dev_addon_info.enabled:
+                output.append(dev_addon_info.path)
 
         return output
 
