@@ -2,8 +2,10 @@ from .exceptions import (
     BundleNotFoundError,
     InstallerDistributionError,
 )
-from .control import AyonDistribution
+from .control import AYONDistribution
 from .utils import (
+    show_missing_permissions,
+    show_blocked_auto_update,
     show_missing_bundle_information,
     show_installer_issue_information,
     UpdateWindowManager,
@@ -14,8 +16,10 @@ __all__ = (
     "BundleNotFoundError",
     "InstallerDistributionError",
 
-    "AyonDistribution",
+    "AYONDistribution",
 
+    "show_missing_permissions",
+    "show_blocked_auto_update",
     "show_missing_bundle_information",
     "show_installer_issue_information",
     "UpdateWindowManager",
