@@ -4,7 +4,7 @@ import sys
 import json
 import subprocess
 
-import appdirs
+import platformdirs
 import semver
 
 IS_BUILT_APPLICATION = getattr(sys, "frozen", False)
@@ -22,7 +22,7 @@ def get_ayon_appdirs(*args):
     """
 
     return os.path.join(
-        appdirs.user_data_dir("AYON", "Ynput"),
+        platformdirs.user_data_dir("AYON", "Ynput"),
         *args
     )
 
